src/presenters/requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pdfkit==0.6.1
PyJWT==1.7.1
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.3
six==1.14.0
Werkzeug==0.16.0
